refactor(provider): keep K3 coverage within lint budgets
Problem:
The initial Kimi K3 change added coverage and branches directly to files that already carry repository lint baselines, causing the incremental repolint check to fail.

Root cause:
New tests lived in oversized shared test files, and protocol validation increased the complexity of existing provider normalization functions.

Fix:
- move Kimi K3 config, boot, and provider tests into focused files
- extract OpenAI and Kimi effort normalization helpers
- isolate Kimi K3 contract detection and effort vocabulary from the main provider constructor
- keep the settings protocol label line-neutral in the existing large component

Verification:
- go run ./tools/repolint
- go test ./...
- go vet ./...
- ./scripts/cache-guard.sh
- pnpm lint:hooks
- pnpm typecheck
- pnpm exec tsx src/__tests__/provider-editor-model-picker.test.tsx
