+++ internal/config/edit_test.go
@@ -2574,24 +2574,6 @@ func TestReasoningProtocolOverrideControlsEffortCapability(t *testing.T) {
 	if got, err := NormalizeEffort(e, "medium"); err != nil || got != "medium" {
 		t.Fatalf("NormalizeEffort(medium) = %q/%v, want medium/nil", got, err)
 	}
-
-	e.ReasoningProtocol = ReasoningProtocolKimiK3
-	cap = EffortCapabilityForEntry(e)
-	wantLevels = []string{"auto", "low", "high", "max"}
-	if !cap.Supported || cap.Default != "max" || !reflect.DeepEqual(cap.Levels, wantLevels) {
-		t.Fatalf("reasoning_protocol=kimi-k3 capability = %+v, want levels %v with default max", cap, wantLevels)
-	}
-	if protocol := ReasoningProtocolForEntry(e); protocol != ReasoningProtocolKimiK3 {
-		t.Fatalf("protocol = %q, want kimi-k3", protocol)
-	}
-	for _, level := range []string{"low", "high", "max"} {
-		if got, err := NormalizeEffort(e, level); err != nil || got != level {
-			t.Fatalf("NormalizeEffort(%s) = %q/%v, want %s/nil", level, got, err, level)
-		}
-	}
-	if _, err := NormalizeEffort(e, "medium"); err == nil {
-		t.Fatal("Kimi K3 reasoning_protocol should reject medium")
-	}
 }
 
 func TestNormalizeEffortCustomSupportedEfforts(t *testing.T) {
@@ -2705,41 +2687,6 @@ func TestUpsertProviderNormalizesCustomEffortFields(t *testing.T) {
 		if got.SupportedEfforts[i] != want {
 			t.Fatalf("supported_efforts[%d] = %q, want %q", i, got.SupportedEfforts[i], want)
 		}
-	}
-}
-
-func TestKimiK3ReasoningProtocolRoundTripsConfig(t *testing.T) {
-	path := filepath.Join(t.TempDir(), "config.toml")
-	c := &Config{}
-	if err := c.UpsertProvider(ProviderEntry{
-		Name:              "custom-kimi-gateway",
-		Kind:              "openai",
-		BaseURL:           "https://gateway.example.com/v1",
-		Model:             "kimi-k3",
-		ReasoningProtocol: " KIMI-K3 ",
-		SupportedEfforts:  []string{"low", "high", "max"},
-		DefaultEffort:     "max",
-	}); err != nil {
-		t.Fatalf("UpsertProvider: %v", err)
-	}
-	if err := c.SaveTo(path); err != nil {
-		t.Fatalf("SaveTo: %v", err)
-	}
-
-	loaded := LoadForEditWithoutCredentials(path)
-	got, ok := loaded.Provider("custom-kimi-gateway")
-	if !ok {
-		t.Fatal("saved Kimi K3 provider was not reloaded")
-	}
-	if got.ReasoningProtocol != ReasoningProtocolKimiK3 || !reflect.DeepEqual(got.SupportedEfforts, []string{"low", "high", "max"}) || got.DefaultEffort != "max" {
-		t.Fatalf("reloaded Kimi K3 provider = %+v", got)
-	}
-	body, err := os.ReadFile(path)
-	if err != nil {
-		t.Fatalf("read saved config: %v", err)
-	}
-	if !strings.Contains(string(body), `reasoning_protocol = "kimi-k3"`) {
-		t.Fatalf("saved config lost Kimi K3 protocol:\n%s", body)
 	}
 }
