+++ internal/provider/openai/kimi_k3.go
@@ -1,0 +1,16 @@
+package openai
+
+import "strings"
+
+var kimiK3EffortVocabulary = []string{"low", "high", "max"}
+
+func usesKimiK3Contract(protocol, baseURL, model string) bool {
+	return protocol == "kimi-k3" || (IsKimiAPI(baseURL) && strings.EqualFold(strings.TrimSpace(model), "kimi-k3"))
+}
+
+func reasoningEffortVocabulary(kimiK3 bool, configured []string) ([]string, bool) {
+	if kimiK3 {
+		return kimiK3EffortVocabulary, true
+	}
+	return configured, hasExplicitSupportedEfforts(configured)
+}
