+++ internal/provider/openai/kimi_k3_test.go
@@ -1,0 +1,39 @@
+package openai
+
+import (
+	"testing"
+
+	"reasonix/internal/provider"
+)
+
+func TestExplicitKimiK3GatewayRequestContract(t *testing.T) {
+	p, err := New(provider.Config{
+		Name: "custom-kimi-gateway", BaseURL: "https://gateway.example.com/v1", Model: "kimi-k3",
+		Extra: map[string]any{"effort": "high", "reasoning_protocol": "kimi-k3"},
+	})
+	if err != nil {
+		t.Fatalf("New explicit Kimi K3 gateway: %v", err)
+	}
+	if !provider.RequiresReasoningRoundTrip(p) {
+		t.Fatal("explicit Kimi K3 protocol must retain reasoning for complete assistant-message replay")
+	}
+	req := p.(*client).buildRequest(provider.Request{
+		Temperature: provider.TemperaturePtr(0.2), MaxTokens: 4096,
+		Messages: []provider.Message{
+			{Role: provider.RoleUser, Content: "question"},
+			{Role: provider.RoleAssistant, Content: "answer", ReasoningContent: "gateway reasoning"},
+		},
+	})
+	if req.ReasoningEffort != "high" || req.Temperature != nil || req.MaxTokens != 0 || req.MaxCompletionTokens != 4096 {
+		t.Fatalf("explicit Kimi K3 request shape = %+v", req)
+	}
+	if got := req.Messages[1].ReasoningContent; got == nil || *got != "gateway reasoning" {
+		t.Fatalf("explicit Kimi K3 reasoning_content = %v, want gateway reasoning", got)
+	}
+	if _, err := New(provider.Config{
+		Name: "custom-kimi-gateway", BaseURL: "https://gateway.example.com/v1", Model: "kimi-k3",
+		Extra: map[string]any{"effort": "medium", "reasoning_protocol": "kimi-k3"},
+	}); err == nil {
+		t.Fatal("explicit Kimi K3 protocol should reject unsupported medium effort")
+	}
+}
