+++ internal/provider/openai/openai_test.go
@@ -1229,48 +1229,6 @@ func TestBuildRequestKimiK3OfficialWireShape(t *testing.T) {
 	if gatewayReq.Temperature == nil || gatewayReq.MaxTokens != 77 || gatewayReq.MaxCompletionTokens != 0 {
 		t.Fatalf("relay request was changed by official Kimi compatibility: %+v", gatewayReq)
 	}
-
-	explicitKimi, err := New(provider.Config{
-		Name:    "custom-kimi-gateway",
-		BaseURL: "https://gateway.example.com/v1",
-		Model:   "kimi-k3",
-		Extra: map[string]any{
-			"effort":             "high",
-			"reasoning_protocol": "kimi-k3",
-		},
-	})
-	if err != nil {
-		t.Fatalf("New explicit Kimi K3 gateway: %v", err)
-	}
-	if !provider.RequiresReasoningRoundTrip(explicitKimi) {
-		t.Fatal("explicit Kimi K3 protocol must retain reasoning for complete assistant-message replay")
-	}
-	explicitReq := explicitKimi.(*client).buildRequest(provider.Request{
-		Temperature: provider.TemperaturePtr(0.2),
-		MaxTokens:   4096,
-		Messages: []provider.Message{
-			{Role: provider.RoleUser, Content: "question"},
-			{Role: provider.RoleAssistant, Content: "answer", ReasoningContent: "gateway reasoning"},
-		},
-	})
-	if explicitReq.ReasoningEffort != "high" || explicitReq.Temperature != nil || explicitReq.MaxTokens != 0 || explicitReq.MaxCompletionTokens != 4096 {
-		t.Fatalf("explicit Kimi K3 request shape = %+v", explicitReq)
-	}
-	if got := explicitReq.Messages[1].ReasoningContent; got == nil || *got != "gateway reasoning" {
-		t.Fatalf("explicit Kimi K3 reasoning_content = %v, want gateway reasoning", got)
-	}
-
-	if _, err := New(provider.Config{
-		Name:    "custom-kimi-gateway",
-		BaseURL: "https://gateway.example.com/v1",
-		Model:   "kimi-k3",
-		Extra: map[string]any{
-			"effort":             "medium",
-			"reasoning_protocol": "kimi-k3",
-		},
-	}); err == nil {
-		t.Fatal("explicit Kimi K3 protocol should reject unsupported medium effort")
-	}
 }
 
 func TestBuildRequestUsesProviderSpecificOutputBudget(t *testing.T) {
